Handwritten digits recognition using Python.
The project has 10,000 scanned handwritten digits as test data. Output is the fraction of digits recgnised correctly after each epoch.
References: http://neuralnetworksanddeeplearning.com/chap1.html
The code given in the link above is in Python version 2. I made some changes to make it compatible with Python 3.6.
The project is a good way to understand the implementation of Machine learning algorithm (here Stoch Gradient Desccent technique is used).
You need to run the driver program file to get the output.